feat: Implement API Key authentication and management
- Add ApiKeyAuthenticationHandler for handling API key authentication.
- Create ApiKeyService for managing API key lifecycle and validation.
- Define ApiKeyAuthInfo model to encapsulate API key authentication information.
- Implement ApiKeyAuthIndexManager for CRUD operations on API keys.
- Introduce DTOs for API key management: ApplicationApiKeyAddDto, ApplicationApiKeyCredentialResultDto, ApplicationApiKeyItemDto.
- Add integration tests for API key CRUD operations and authorization checks.

+++ src/Modules/ModelMod/Models/ApplicationQuotaDtos/ApplicationQuotaAddDto.cs
@@ -12,7 +12,7 @@ public class ApplicationQuotaAddDto
 
     public int MaxRequests { get; set; }
 
-    public int MaxTokens { get; set; }
+    public long MaxTokens { get; set; }
 
     public int WindowSeconds { get; set; }
