+++ src/Modules/ModelMod/Models/ApplicationQuotaDtos/ApplicationQuotaDetailDto.cs
@@ -10,7 +10,7 @@ public class ApplicationQuotaDetailDto
     public Guid ApplicationId { get; set; }
     public QuotaPeriodType PeriodType { get; set; }
     public int MaxRequests { get; set; }
-    public int MaxTokens { get; set; }
+    public long MaxTokens { get; set; }
     public int WindowSeconds { get; set; }
     public bool IsEnabled { get; set; }
     public DateTimeOffset CreatedTime { get; set; }
