+++ src/Modules/ModelMod/Models/ApplicationQuotaDtos/ApplicationQuotaItemDto.cs
@@ -10,7 +10,7 @@ public class ApplicationQuotaItemDto
     public Guid ApplicationId { get; set; }
     public QuotaPeriodType PeriodType { get; set; }
     public int MaxRequests { get; set; }
-    public int MaxTokens { get; set; }
+    public long MaxTokens { get; set; }
     public int WindowSeconds { get; set; }
     public bool IsEnabled { get; set; }
 }
