+++ src/Modules/ModelMod/Models/ApplicationQuotaDtos/ApplicationQuotaUpdateDto.cs
@@ -10,7 +10,7 @@ public class ApplicationQuotaUpdateDto
 
     public int? MaxRequests { get; set; }
 
-    public int? MaxTokens { get; set; }
+    public long? MaxTokens { get; set; }
 
     public int? WindowSeconds { get; set; }
